% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A numeric scalar}

\item{y}{A numeric scalar}
}
\value{
The highest common divisor of \code{x} and \code{y}
}
\description{
Gives the highest common divisor of two integers.
}
\details{
Uses the Euclidean algorithm to find the greatest common divisor between two numbers.
}
\examples{
euclidean(700, 20)
euclidean(-5552, 1515126)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
