\name{sayBye}
\alias{sayBye}
\title{
a sayHello function demo
}
\description{
a sayHello function demo
}
\usage{
sayBye(name)
}
\arguments{
  \item{name}{
a word}
}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Yu
}
\note{
nothing
}

\seealso{
nothing
}
\examples{
sayBye('body')
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ sayBye }% use one of  RShowDoc("KEYWORDS")
