% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayhello.R
\name{sayhello}
\alias{sayhello}
\title{Say Hello!}
\usage{
sayhello(name = NULL)
}
\arguments{
\item{name}{character; the name you would like to print hello to}
}
\value{
the supplied name invisibly
}
\description{
This function will print "Hello!" with a name if included.
}
\examples{
sayhello()
sayhello("Samantha")
sayhello("Chris")
}
