% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayHelloData.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{A vertex in the graph}
  \item{v2}{A vertex in the graph}
  \item{w}{The weight of the edge described by connecting v1 and v2}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A data frame describing the first graph shown on the Wikipedia page for Dijkstra's algorithm.
}
\keyword{datasets}
