\name{sayHello}
\alias{sayHello}
\title{Say Hello function}
\description{
  This is a demo funciton for building a package.
}
\usage{
  sayHello(name)
}
\arguments{
  \item{name}{
  a word
}
}
\details{
  no details
}
\value{
  no return
}
\references{
  \url{http://blog.fens.me/r-build-package}
}
\author{
  Dan Zhang <\email{bsspirit@gmail.com}>
}
\note{
  nothing
}

\seealso{
  nothing
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## Sample
sayHello("abc")
}
\keyword{ sayHello }
