% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayHello.R
\name{sayHello}
\alias{sayHello}
\title{sayHello}
\usage{
sayHello(name)
}
\arguments{
\item{name}{name is a word}
}
\value{
a word
}
\description{
the function paste two words.
}
\details{
This is some description

you can use this function.
}
\examples{
name="FUCHAO";sayHello(name)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
